+++ src/main/java/de/embl/cba/mobie/bdv/MobieBdvSupplier.java
@@ -37,7 +37,6 @@ public class MobieBdvSupplier implements IBdvSupplier {
         Prefs.sourceNameOverlayPosition( Prefs.OverlayPosition.TOP_RIGHT );
         
         BdvStackSource<ByteType> bss = BdvFunctions.show( dummyImg, "dummy", options );
-        bss.getSources().get( 0 ).getSpimSource()
         BdvHandle bdv = bss.getBdvHandle();
 
         if ( sOptions.interpolate ) bdv.getViewerPanel().setInterpolation( Interpolation.NLINEAR );
