+++ src/main/java/de/embl/cba/mobie/bdv/view/SegmentationSliceView.java
@@ -5,6 +5,7 @@ import bdv.util.BdvHandle;
 import bdv.viewer.SourceAndConverter;
 import bdv.viewer.TimePointListener;
 import de.embl.cba.mobie.MoBIE;
+import de.embl.cba.mobie.Utils;
 import de.embl.cba.mobie.bdv.render.BlendingMode;
 import de.embl.cba.mobie.color.OpacityAdjuster;
 import de.embl.cba.mobie.n5.source.LabelSource;
@@ -15,6 +16,7 @@ import de.embl.cba.mobie.transform.TransformHelper;
 import de.embl.cba.tables.color.ColoringListener;
 import de.embl.cba.tables.imagesegment.ImageSegment;
 import de.embl.cba.tables.select.SelectionListener;
+import mpicbg.spim.data.sequence.VoxelDimensions;
 import net.imglib2.realtransform.AffineTransform3D;
 import sc.fiji.bdvpg.bdv.BdvHandleHelper;
 import sc.fiji.bdvpg.bdv.navigate.ViewerTransformChanger;
@@ -25,6 +27,7 @@ import javax.swing.*;
 import java.awt.*;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class SegmentationSliceView< S extends ImageSegment > implements ColoringListener, SelectionListener< S >
 {
@@ -48,15 +51,12 @@ public class SegmentationSliceView< S extends ImageSegment > implements Coloring
 		display.selectionModel.listeners().add( this );
 		display.coloringModel.listeners().add( this );
 
-		// open
-		List< SourceAndConverter< ? > > sourceAndConverters = moBIE.openSourceAndConverters( display.getSources() );
-
-		// transform
-		sourceAndConverters = TransformHelper.transformSourceAndConverters( sourceAndConverters, display.sourceTransformers );
+		List< SourceAndConverter< ? > > sourceAndConverters = display.getSources().stream().map( name -> moBIE.getSourceAndConverter( name ) ).collect( Collectors.toList() );
 
 		// convert to labelSource
 		sourceAndConverters = asLabelSources( sourceAndConverters );
 
+		display.sourceAndConverters = new ArrayList<>();
 		for ( SourceAndConverter< ? > sourceAndConverter : sourceAndConverters )
 		{
 			// set opacity
@@ -70,9 +70,9 @@ public class SegmentationSliceView< S extends ImageSegment > implements Coloring
 				SourceAndConverterServices.getSourceAndConverterService().setMetadata( sourceAndConverter, BlendingMode.BLENDING_MODE, display.getBlendingMode() );
 
 			bdvHandle.getViewerPanel().addTimePointListener( ( TimePointListener ) sourceAndConverter.getConverter() );
-		}
 
-		display.sourceAndConverters = sourceAndConverters;
+			display.sourceAndConverters.add( sourceAndConverter );
+		}
 	}
 
 	private List< SourceAndConverter< ? > > asLabelSources( List< SourceAndConverter< ? > > sourceAndConverters )
@@ -149,22 +149,19 @@ public class SegmentationSliceView< S extends ImageSegment > implements Coloring
 
 	private void adaptPosition( double[] position, String sourceName )
 	{
-		// TODO: simply check where the source with sourceName sits right now!
-		//   one has to remove the voxelSize to adapt the position
-		// remove from the transformers the getTransform( ) method
-		if ( display.sourceTransformers != null )
-		{
-			for ( SourceTransformer sourceTransformer : display.sourceTransformers )
-			{
-				final AffineTransform3D transform = sourceTransformer.getTransform( sourceName );
-				if ( transform != null )
-				{
-					// not each transformer of this display may transform all sources
-					// this a transform can be null
-					transform.apply( position, position );
-				}
-			}
-		}
+		// get source transform
+		final SourceAndConverter< ? > sourceAndConverter = Utils.getSourceAndConverter( display.sourceAndConverters, sourceName );
+		AffineTransform3D sourceTransform = new AffineTransform3D();
+		sourceAndConverter.getSpimSource().getSourceTransform( 0,0, sourceTransform );
+
+		// remove scaling, because the positions are in scaled units
+		final VoxelDimensions voxelDimensions = sourceAndConverter.getSpimSource().getVoxelDimensions();
+		final AffineTransform3D scalingTransform = new AffineTransform3D();
+		scalingTransform.scale( voxelDimensions.dimension( 0 ), voxelDimensions.dimension( 1 ), voxelDimensions.dimension( 2 )  );
+		sourceTransform = sourceTransform.concatenate( scalingTransform.inverse() );
+
+		// adapt
+		sourceTransform.apply( position, position );
 	}
 
 	public BdvHandle getBdvHandle()
