+++ src/main/java/de/embl/cba/mobie/transform/TransformHelper.java
@@ -14,20 +14,6 @@ import java.util.List;
 
 public class TransformHelper
 {
-	public static List< SourceAndConverter< ? > > transformSourceAndConverters( List< SourceAndConverter< ? > > sourceAndConverters, List< SourceTransformer > sourceTransformers )
-	{
-		if ( sourceTransformers != null )
-		{
-			for ( SourceTransformer sourceTransformer : sourceTransformers )
-			{
-				sourceAndConverters = sourceTransformer.transform( sourceAndConverters );
-				//break;
-			}
-		}
-
-		return sourceAndConverters;
-	}
-
 	public static RealInterval unionRealInterval( List< ? extends Source< ? > > sources )
 	{
 		RealInterval union = null;
